Add Villager Basics
Create villager_demo scene for prototyping villager vehavior.

Add VillageCenter object to handle villager creation and management. Should also be used to define a village and its boundaries

Add basics of villager movement and start on the AI. With some refactoring a lot of it could be combined with the boat AI. Just supports GoTo, Random, and Wait for now. Has no awareness of paths or objectives.